Unify project, add cross-platform GUI & real emitters
Refactored solution to a single project with platform-gated GUI hosts for Windows (Win32) and macOS (Cocoa), using conditional compilation via MSBuild symbols. Added new binary emitters for Mach-O arm64 and Win32 PE (x86/x64), with minimal assembler backends. Centralized CLI, emitters, and compilation workflow. Updated documentation and CLI help to reflect unified architecture. All code now lives in one project, simplifying build and maintenance.
