--- BrainFudger.Core.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <AssemblyName>BrainFudger.Core</AssemblyName>
-    <RootNamespace>BrainFudger</RootNamespace>
-    <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
-    <InvariantGlobalization>true</InvariantGlobalization>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="Emitters\**\*.cs" />
-    <Compile Include="Models\**\*.cs" />
-    <Compile Include="Services\**\*.cs" />
-  </ItemGroup>
-</Project>
