--- BrainFudger.Mac/BrainFudger.Mac.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <AssemblyName>BrainFudger.Mac</AssemblyName>
-    <RootNamespace>BrainFudger.Mac</RootNamespace>
-    <InvariantGlobalization>true</InvariantGlobalization>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\BrainFudger.Core.csproj" />
-  </ItemGroup>
-</Project>
