--- BrainFudger.sln
@@ -1,33 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BrainFudger", "BrainFudger.csproj", "{7D2C3EC3-8E45-4F21-BB75-E9A0C7FD4F1C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BrainFudger.Core", "BrainFudger.Core.csproj", "{DCEB48CC-919F-42A6-85D0-88461C7B3957}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BrainFudger.Mac", "BrainFudger.Mac\BrainFudger.Mac.csproj", "{88602B17-FD20-45D6-B3E4-B8FF2BC47FCF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7D2C3EC3-8E45-4F21-BB75-E9A0C7FD4F1C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7D2C3EC3-8E45-4F21-BB75-E9A0C7FD4F1C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7D2C3EC3-8E45-4F21-BB75-E9A0C7FD4F1C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7D2C3EC3-8E45-4F21-BB75-E9A0C7FD4F1C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DCEB48CC-919F-42A6-85D0-88461C7B3957}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DCEB48CC-919F-42A6-85D0-88461C7B3957}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DCEB48CC-919F-42A6-85D0-88461C7B3957}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DCEB48CC-919F-42A6-85D0-88461C7B3957}.Release|Any CPU.Build.0 = Release|Any CPU
-		{88602B17-FD20-45D6-B3E4-B8FF2BC47FCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{88602B17-FD20-45D6-B3E4-B8FF2BC47FCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{88602B17-FD20-45D6-B3E4-B8FF2BC47FCF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{88602B17-FD20-45D6-B3E4-B8FF2BC47FCF}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
