+++ BrainFudger/Emitters/BinaryEmitters.cs
