+++ BrainFudger/Emitters/DosEmitter.cs
