+++ BrainFudger/Emitters/MachOEmitter.cs
