+++ BrainFudger/Emitters/PortableExecutableEmitter.cs
