+++ BrainFudger/Gui/IGuiApplicationHost.cs
@@ -1,0 +1,9 @@
+﻿namespace BrainFudger.Gui;
+
+internal interface IGuiApplicationHost
+{
+    bool TryHandleHostArguments(string[] args, out int exitCode);
+    bool ShouldLaunchDetached();
+    int LaunchDetached();
+    int Run();
+}
